+++ src/components/Modal/Modal.tsx
@@ -70,6 +70,7 @@ const Modal = (props: T.Props) => {
 		autoFocus = true,
 		disableSwipeGesture,
 		overlayClassName,
+		containerRef,
 		className,
 		attributes,
 	} = props;
@@ -231,6 +232,7 @@ const Modal = (props: T.Props) => {
 			active={active}
 			transparent={transparentOverlay || hideProgress}
 			className={overlayClassName}
+			containerRef={containerRef}
 			attributes={{
 				onTouchStart: handleDragStart,
 			}}