+++ src/components/_private/Portal/Portal.types.ts
@@ -2,7 +2,7 @@ import React from "react";
 
 export type Props = {
 	children?: React.ReactNode;
-	targetRef?: React.RefObject<HTMLElement | null>;
+	targetRef?: React.RefObject<HTMLElement | ShadowRoot | null>;
 };
 
 export type ScopeProps<T extends HTMLElement> = {